poc_implementation/requirements.txt
llamaindex>=0.10.0
python-docx
pymupdf
docxtpl
sentence-transformers
torch
numpy
pandas
camel-tools
transformers
arabic-reshaper
python-bidi
qdrant-client
langchain
langchain-community
langgraph
openai
tiktoken
datasets
evaluate
accelerate
bitsandbytes
peft
gradio